refactor: address Gemini review
